+++ test/leaf_test.exs
@@ -4,8 +4,4 @@ defmodule LeafTest do
   test "Leaf module is loaded" do
     assert Code.ensure_loaded?(Leaf)
   end
-
-  test "Leaf.Icon module is loaded" do
-    assert Code.ensure_loaded?(Leaf.Icon)
-  end
 end